requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
loguru
python-multipart
aiofiles
celery[redis]
redis
faster-whisper
pyannote.audio
torch
openai
pgvector